Added the logger file

+++ src/main/java/pk/Percentage.java
@@ -2,8 +2,8 @@ package pk;
 
 public class Percentage {
     // Calculates the stdout after the simulation is done
-    public static int stdout(int play, int games){
-        int num = (play/games)*100;
+    public static double stdout(double play, double games){
+        double num = (play/games)*100;
         return num;
     }
 }
